backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-doc
annotated-types
anthropic
anyio
attrs
certifi
cffi
charset-normalizer
click
colorama
cryptography
dataclasses-json
distro
dnspython
docstring_parser
ecdsa
email-validator
fastapi
filelock
filetype
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
hf-xet
httpcore
httplib2
httpx
httpx-sse
huggingface_hub
idna
Jinja2
jiter
joblib
jsonpatch
jsonpointer
langchain
langchain-anthropic
langchain-classic
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
motor
mpmath
multidict
mypy_extensions
networkx
numpy
orjson
ormsgpack
packaging
pillow
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
pymongo
pyparsing
python-dotenv
python-jose
PyYAML
regex
requests
requests-toolbelt
rich
rsa
safetensors
scikit-learn
scipy
sentence-transformers
shellingham
six
sniffio
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tokenizers
torch
torchvision
tqdm
transformers
typer
typer-slim
typing-inspect
typing-inspection
typing_extensions
uritemplate
urllib3
uuid_utils
uvicorn
websockets
xxhash
yarl
zstandard